# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: OpenSeaFeed API
  version: "1.0"
  description: |
    The open, community-owned AIS network. All HTTP endpoints live on
    `https://api.openseafeed.com`; the live WebSocket stream lives on
    `wss://stream.openseafeed.com/v1/stream`; raw NMEA contribution goes to
    `ingest.openseafeed.com`.

    ## Keys and tiers

    Keys are created on the [dashboard](https://api.openseafeed.com/) after
    signing in with GitHub or an email sign-in link.

    | kind | prefix | purpose |
    |---|---|---|
    | live | `osf_live_...` | consuming: stream, snapshot, history |
    | feed | `osf_feed_...` | contributing an upstream feed |
    | station | `osf_stn_...` | contributing your own receiver (via station registration) |

    **Anonymous access is welcome** at the free tier: no key needed for the
    stream (bounding boxes up to 30000 square degrees of area), single-vessel state,
    recent history (last 48 h, downsampled) and the delayed fleet snapshot.
    Accounts that actively contribute a feed or receiver get the
    `contributor` tier: unlimited stream area, the freshest snapshot, and
    unclamped history.

    ## Getting data: the WebSocket stream

    Connect to `wss://stream.openseafeed.com/v1/stream` and send one JSON
    subscription within 3 seconds (aisstream.io-compatible; resend at any
    time to swap the subscription):

    ```json
    {
      "APIKey": "osf_live_...",
      "BoundingBoxes": [[[54, -5], [72, 32]]],
...
    }
    ```

...

...

...

...

...

...

...
servers:
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: key
      description: An `osf_live_...` key. Omit entirely for anonymous free-tier access.
    sessionCookie:
      type: apiKey
      in: cookie
      name: osf_session
      description: Signed session cookie set by the sign-in flows.
  schemas:
...
    HistoryPoint:
      type: object
      properties:
        ts: { type: string, example: "2026-08-17 09:40:00" }
        msg_type: { type: string, example: PositionReport }
        lat: { type: number }
        lon: { type: number }
        sog: { type: number, nullable: true }
        cog: { type: number, nullable: true }
        heading: { type: integer, nullable: true }
        nav_status: { type: integer, nullable: true }
...
paths:
...
  /v1/history/{mmsi}:
    get:
      tags: [vessels]
      summary: Position history of one vessel
      description: |
        Stored track from the archive. Anonymous requests are clamped to the
        last 48 hours and always downsampled (step >= 30 s); a valid key
        unlocks the full range (five-year design horizon).
      security: [{}, { apiKey: [] }]
      parameters:
        - { name: mmsi, in: path, required: true, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string, format: date-time }, description: "Default: to - 24 h" }
        - { name: to, in: query, schema: { type: string, format: date-time }, description: "Default: now" }
        - { name: step, in: query, schema: { type: integer, minimum: 10 }, description: Downsample to one point per step seconds
          }
        - { name: limit, in: query, schema: { type: integer, maximum: 50000 } }
      responses:
        "200":
          description: Track
          content:
            application/json:
              schema:
                type: object
                properties:
                  mmsi: { type: integer }
                  from: { type: string, format: date-time }
                  to: { type: string, format: date-time }
                  count: { type: integer }
                  points:
                    type: array
                    items: { $ref: "#/components/schemas/HistoryPoint" }
...